Create listener package and add javadoc
