--- libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/FeedbackListener.java
@@ -1,11 +1,0 @@
-package com.mapbox.services.android.navigation.ui.v5;
-
-import com.mapbox.services.android.navigation.ui.v5.feedback.FeedbackItem;
-
-public interface FeedbackListener {
-  void onFeedbackOpened();
-
-  void onFeedbackCancelled();
-
-  void onFeedbackSent(FeedbackItem feedbackItem);
-}
