--- libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationListener.java
@@ -1,10 +1,0 @@
-package com.mapbox.services.android.navigation.ui.v5;
-
-public interface NavigationListener {
-
-  void onCancelNavigation();
-
-  void onNavigationFinished();
-
-  void onNavigationRunning();
-}
